fix : api 객체 형태 변경
