+++ src/constant/cookie.ts
@@ -1,0 +1,4 @@
+export const COOKIE_KEYS = {
+  accessToken: 'swifty-access',
+  refreshToken: 'swifty-refresh',
+} as const;
