+++ tsconfig.json
@@ -20,7 +20,8 @@
       }
     ],
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@constant": ["./src/constant"]
     }
   },
   "include": [
